# Optimize cilium/ebpf kallsyms: 27% initialization speedup for trace_dns