@alban 2025-06-25
Replaced by:
- https://github.com/inspektor-gadget/inspektor-gadget/pull/4653
- https://github.com/inspektor-gadget/inspektor-gadget/pull/4652

See also discussion in:
- https://github.com/cilium/ebpf/pull/1802#issuecomment-3000806170